add varkes-config submodule
