+++ modules/config-validator/.npmignore
@@ -1,0 +1,4 @@
+node_modules
+Makefile
+*.log
+src
