+++ modules/odata-mock/package.json
@@ -18,6 +18,7 @@
     "access": "public"
   },
   "dependencies": {
+    "@varkes/config-validator": "0.0.1",
     "body-parser": "^1.19.0",
     "compression": "^1.7.4",
     "cors": "^2.8.5",
