+++ modules/odata-mock/src/test/test.ts
@@ -1,13 +1,13 @@
 #!/usr/bin/env node
 'use strict'
 
-import { init as mock } from "../server/app"
+import * as mock from "../server/app"
 const request = require('supertest')
 const express = require('express')
 
 describe('test app', function () {
     it('should work', (done) => {
-        mock('./varkes_config.json', __dirname).then((mock: any) => {
+        mock.init('varkes_config.json', __dirname).then((mock: any) => {
             let app = express()
             app.use(mock)
