+++ modules/openapi-mock/package.json
@@ -11,6 +11,7 @@
     "access": "public"
   },
   "dependencies": {
+    "@varkes/config-validator": "0.0.1",
     "api-spec-converter": "~2.9.1",
     "body-parser": "^1.19.0",
     "express": "^4.17.1",
